backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via django
django==5.0
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.14.0
    # via -r requirements.in
psycopg[c]==3.1.16
    # via
    #   -r requirements.in
    #   psycopg
psycopg-c==3.1.16
    # via psycopg
pytz==2023.3.post1
    # via djangorestframework
sqlparse==0.4.4
    # via django
typing-extensions==4.9.0
    # via psycopg
